+++ python/GafferUITest/__init__.py
@@ -130,6 +130,7 @@ from .ToolTest import ToolTest
 from .StandardNodeToolbarTest import StandardNodeToolbarTest
 from .LabelPlugValueWidgetTest import LabelPlugValueWidgetTest
 from .PythonEditorTest import PythonEditorTest
+from .BoxIOUITest import BoxIOUITest
 
 if __name__ == "__main__":
 	unittest.main()